Create the `diskUtil` module
This module contains some convenience helpers for interacting with files
on disk, creating directories, etc. They are factored out of
`cli/common` (we're going to start needing them in the `api` subdir
too).

Test plan: Existing tests ported, and wrote some tests for `mkdirx`.
